+++ wetterdienst/download/download_services.py
@@ -1,6 +1,7 @@
 """
 **DWD download utilities**
 """
+import logging
 from io import BytesIO
 from pathlib import PurePosixPath
 from typing import Union
@@ -8,6 +9,8 @@ from typing import Union
 from wetterdienst.constants.access_credentials import DWDCDCBase
 from wetterdienst.download.https_handling import create_dwd_session
 from wetterdienst.file_path_handling.path_handling import build_dwd_cdc_data_path
+
+logger = logging.getLogger(__name__)
 
 
 def download_file_from_dwd(
@@ -24,7 +27,9 @@ def download_file_from_dwd(
     """
     dwd_session = create_dwd_session()
 
-    r = dwd_session.get(build_dwd_cdc_data_path(filepath, cdc_base))
+    url = build_dwd_cdc_data_path(filepath, cdc_base)
+    logger.info(f"Downloading resource {url}")
+    r = dwd_session.get(url)
     r.raise_for_status()
 
     return BytesIO(r.content)